Ui In game
-Jauge Decontamination
-add asset ui in game

+++ Assets/script/TileData.cs
@@ -6,9 +6,8 @@ public class TileData : MonoBehaviour
     public int ID;
     public bool IsOccupied;
     public GameObject OnTop;
-    public float ProgressValue = 0;
+    public int ProgressValue = 0;
     public ChunkState ChunkState;
-
     //private void OnMouseEnter()
     //{
